requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
